+++ emulators/xfce/uninstall.sh
@@ -2,7 +2,7 @@
 set -e
 
 TARGET_DIR="$HOME/.local/share/xfce4/terminal/colorschemes"
-THEMES_FILES="x.theme madrid.theme lahabana.theme seul.theme miami.theme paris.theme tokio.theme oslo.theme helsinki.theme berlin.theme london.theme praha.theme bogota.theme"
+THEMES_FILES="x.theme madrid.theme lahabana.theme miami.theme paris.theme tokio.theme oslo.theme helsinki.theme berlin.theme london.theme praha.theme bogota.theme"
 
 detect_pm() {
   for pm in apt-get dnf pacman zypper yum apk brew; do